#  Implement the Cohen-Hubbard-Oesterwinter method