--- a/integrators/symmetric_linear_multistep_integrator.hpp
+++ b/integrators/symmetric_linear_multistep_integrator.hpp
@@ -14,6 +14,7 @@
 
 #include "base/status.hpp"
 #include "integrators/backward_difference.hpp"
+#include "integrators/cohen_hubbard_oesterwinter.hpp"
 #include "integrators/ordinary_differential_equations.hpp"
 #include "numerics/double_precision.hpp"
 #include "numerics/fixed_arrays.hpp"
@@ -86,9 +87,13 @@ class SymmetricLinearMultistepIntegrator
     // updated more frequently than once every |instance.step_|.
     void StartupSolve(Instant const& t_final);
 
-    // Performs the velocity computation, i.e. a backward difference using the
-    // positions computed by the main integrator.
-    void ComputeVelocity(int dimension);
+    // Performs the velocity computation using a backward difference based on
+    // the positions computed by the main integrator.
+    void ComputeVelocityUsingBackwardDifferences();
+
+    // Performs the velocity computation using the Cohen-Hubbard-Oesterwinter
+    // method based on the accelerations computed by the main integrator.
+    void ComputeVelocityUsingCohenHubbardOesterwinter();
 
     static void FillStepFromSystemState(ODE const& equation,
                                         typename ODE::SystemState const& state,
@@ -121,6 +126,7 @@ class SymmetricLinearMultistepIntegrator
 
   FixedStepSizeIntegrator<ODE> const& startup_integrator_;
   BackwardDifference<order_ - 1> const& backward_difference_;
+  CohenHubbardOesterwinter<order_ + 1> const& cohen_hubbard_oesterwinter_;
   FixedVector<double, half_order_> const ɑ_;
   FixedVector<double, half_order_> const β_numerator_;
   double const β_denominator_;