--- a/physics/ephemeris_test.cpp
+++ b/physics/ephemeris_test.cpp
@@ -499,9 +499,9 @@ TEST_P(EphemerisTest, EarthProbe) {
   // The solution is a line, so the rounding errors dominate.  Different
   // libms result in different errors and thus different numbers of steps.
   EXPECT_THAT(probe_positions.size(),
-              AnyOf(Eq(395), Eq(414), Eq(438), Eq(450)));
+              AnyOf(Eq(395), Eq(414), Eq(434), Eq(438), Eq(450)));
   EXPECT_THAT(probe_positions.back().coordinates().x,
-              AlmostEquals(1.00 * period * v_probe, 244, 259));
+              AlmostEquals(1.00 * period * v_probe, 230, 259));
   EXPECT_THAT(probe_positions.back().coordinates().y,
               Eq(q_probe));