This brings significant improvements in accuracy and in speed:
Before:
```
---------------------------------------------------------------------------------
Benchmark                                          Time           CPU Iterations
---------------------------------------------------------------------------------
BM_EphemerisSolarSystemMajorBodiesOnly/-3 90719773589 ns 90371379300 ns          1 +9.92179030813603147e-01 ua
```
After:
```
---------------------------------------------------------------------------------
Benchmark                                          Time           CPU Iterations
---------------------------------------------------------------------------------
BM_EphemerisSolarSystemMajorBodiesOnly/-3 55156286341 ns 54912352000 ns          1 +9.92179030813603147e-01 ua
```